test: gotestsum for cleaner summary
gotestsum presents go test results in a prettier
format. This is helpful to see an overview of status.
Currently this uses a small shell script to manage the
package names and pipes through go tool test2json.
In a follow-up maybe we can can add -json to tinygo test
instead.

There are some new temporary makefile targets to show
usage and make it easy to run for now. "-fast" runs the
TEST_PACKAGES_FAST list, "-gap" runs the known list
of failures from that collection only.

  make tinygo-test-wasip2-sum-fast
  make tinygo-test-wasip2-sum-gap